switched all files to LF instead of CRLF. Linux compatible, should avoid files changed issue (where files showed as changed but weren't)
