--- simcon_project/users/tables/__init__.py
@@ -1,1 +1,0 @@
-from .student_home_table import StudentHomeTable
