+++ simcon_project/users/views/student_home.py
@@ -1,8 +1,15 @@
 from django.shortcuts import render
 from django.contrib.auth.decorators import login_required
 from users.models import *
-from users.tables import *
 
+
+class StudentHomeTable(tables.Table):
+    name = tables.Column(linkify={"viewname": "TemplateStartView", "args": [tables.A("conversation_templates__name")]},
+                         accessor='conversation_templates__name',
+                         verbose_name='Template Name')
+    date_assigned = tables.Column(verbose_name='Date Assigned')
+    completion_date = tables.Column(accessor='conversation_templates__template_responses__completion_date',
+                                    verbose_name='Last Response')
 
 @login_required
 def StudentView(request):
